Add User Controller, controller Helper and Auth middleware to usingDb. Add a new function to create users table to db.js and I also add a new one to many relationship between users and reflections table. I update usingDB Reflection controller queries to only query reflections table based on the current userId

+++ src/usingDB/controller/Reflection.js
@@ -11,14 +11,15 @@ const Reflection = {
    */
   async create(req, res) {
     const createQuery = `INSERT INTO
-      reflections(id, success, low_point, take_away, created_date, modified_date)
-      VALUES($1, $2, $3, $4, $5, $6)
+      reflections(id, success, low_point, take_away, owner_id, created_date, modified_date)
+      VALUES($1, $2, $3, $4, $5, $6, $7)
       returning *`;
     const values = [
       uuidv4(),
       req.body.success,
       req.body.low_point,
       req.body.take_away,
+      req.user.id,
       moment(new Date()),
       moment(new Date())
     ];
@@ -31,15 +32,15 @@ const Reflection = {
     }
   },
   /**
-   * Get All Reflection
+   * Get All Reflections
    * @param {object} req 
    * @param {object} res 
    * @returns {object} reflections array
    */
   async getAll(req, res) {
-    const findAllQuery = 'SELECT * FROM reflections';
+    const findAllQuery = 'SELECT * FROM reflections where owner_id = $1';
     try {
-      const { rows, rowCount } = await db.query(findAllQuery);
+      const { rows, rowCount } = await db.query(findAllQuery, [req.user.id]);
       return res.status(200).send({ rows, rowCount });
     } catch(error) {
       return res.status(400).send(error);
@@ -52,9 +53,9 @@ const Reflection = {
    * @returns {object} reflection object
    */
   async getOne(req, res) {
-    const text = 'SELECT * FROM reflections WHERE id = $1';
+    const text = 'SELECT * FROM reflections WHERE id = $1 AND owner_id = $2';
     try {
-      const { rows } = await db.query(text, [req.params.id]);
+      const { rows } = await db.query(text, [req.params.id, req.user.id]);
       if (!rows[0]) {
         return res.status(404).send({'message': 'reflection not found'});
       }
@@ -70,12 +71,12 @@ const Reflection = {
    * @returns {object} updated reflection
    */
   async update(req, res) {
-    const findOneQuery = 'SELECT * FROM reflections WHERE id=$1';
+    const findOneQuery = 'SELECT * FROM reflections WHERE id=$1 AND owner_id = $2';
     const updateOneQuery =`UPDATE reflections
       SET success=$1,low_point=$2,take_away=$3,modified_date=$4
-      WHERE id=$5 returning *`;
+      WHERE id=$5 AND owner_id = $6 returning *`;
     try {
-      const { rows } = await db.query(findOneQuery, [req.params.id]);
+      const { rows } = await db.query(findOneQuery, [req.params.id, req.user.id]);
       if(!rows[0]) {
         return res.status(404).send({'message': 'reflection not found'});
       }
@@ -84,7 +85,8 @@ const Reflection = {
         req.body.low_point || rows[0].low_point,
         req.body.take_away || rows[0].take_away,
         moment(new Date()),
-        req.params.id
+        req.params.id,
+        req.user.id
       ];
       const response = await db.query(updateOneQuery, values);
       return res.status(200).send(response.rows[0]);
@@ -99,9 +101,9 @@ const Reflection = {
    * @returns {void} return statuc code 204 
    */
   async delete(req, res) {
-    const deleteQuery = 'DELETE FROM reflections WHERE id=$1 returning *';
+    const deleteQuery = 'DELETE FROM reflections WHERE id=$1 AND owner_id = $2 returning *';
     try {
-      const { rows } = await db.query(deleteQuery, [req.params.id]);
+      const { rows } = await db.query(deleteQuery, [req.params.id, req.user.id]);
       if(!rows[0]) {
         return res.status(404).send({'message': 'reflection not found'});
       }
